requirements.txt
drake
numpy
gymnasium
stable-baselines3
matplotlib  # optional
